# Conflicts:
#	.gitignore
#	src/main/java/com/example/paymentservice/persistence/model/Payment.java
#	src/main/java/com/example/paymentservice/web/dto/PaymentDto.java
#	src/main/java/com/example/paymentservice/web/dto/ResponseDto.java
#	src/main/resources/application.yaml
#	src/test/java/com/example/paymentservice/app/service/PaymentServiceTest.java
#	src/test/java/com/example/paymentservice/web/controller/PaymentControllerTest.java

+++ src/main/java/com/example/paymentservice/persistence/model/Payment.java
@@ -2,7 +2,6 @@ package com.example.paymentservice.persistence.model;
 
 import jakarta.persistence.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -15,7 +14,7 @@ public class Payment {
     private String customerName;
 
     @Column(name = "payment_amount", nullable = false)
-    private BigDecimal paymentAmount;
+    private Integer paymentAmount;
 
     @Column(name = "payment_date", nullable = false)
     @Temporal(TemporalType.TIMESTAMP)
@@ -24,7 +23,7 @@ public class Payment {
     public Payment() {
     }
 
-    public Payment(Long orderId, String customerName, BigDecimal paymentAmount, LocalDateTime paymentDate) {
+    public Payment(Long orderId, String customerName, Integer paymentAmount, LocalDateTime paymentDate) {
         this.orderId = orderId;
         this.customerName = customerName;
         this.paymentAmount = paymentAmount;
@@ -47,11 +46,11 @@ public class Payment {
         this.customerName = customerName;
     }
 
-    public BigDecimal getPaymentAmount() {
+    public Integer getPaymentAmount() {
         return paymentAmount;
     }
 
-    public void setPaymentAmount(BigDecimal paymentAmount) {
+    public void setPaymentAmount(Integer paymentAmount) {
         this.paymentAmount = paymentAmount;
     }
